SCRUM-16: refactor DatePicker, Calendar, Popover, styles, stories
