--- src/common/components/calendar/calendar.module.css
@@ -1,6 +1,0 @@
-.wrapper {
-  position: absolute;
-  left: -79px;
-  top: -5px;
-  max-width: 300px;
-}
